TCGA LUSC gene expression subset (Ensembl IDs, log2 TPM)

TCGA LUSC gene expression subset (Ensembl IDs, log2 TPM)

data(TCGA_LUSC_ENSEMBL)

Format

A numeric matrix with 22,962 rows (Ensembl IDs) and 50 columns (primary tumor samples):

rownames

Ensembl gene identifiers (e.g., ENSG00000141510).

colnames

TCGA sample barcodes (e.g., TCGA-XX-XXXX-01A).

Source

Derived from TCGA_LUSC (see ?TCGA_LUSC). The original data were obtained from https://portal.gdc.cancer.gov/projects/TCGA-LUSC.

Details

The matrix is identical of TCGA_LUSC, with rownames converted to Ensembl IDs using AnnotationDbi::mapIds with keytype = "SYMBOL" and column = "ENSEMBL". Expression values are identical to the corresponding rows of TCGA_LUSC (log2(TPM + 1)). This dataset is intended to demonstrate the idType = "ENSEMBL" functionality of the classifyHNSC() function.

This dataset is used in the vignette and examples to show how classifyHNSC() handles non-symbol input and performs automatic ID conversion.

Examples

data(TCGA_LUSC_ENSEMBL)
dim(TCGA_LUSC_ENSEMBL)
#> [1] 22962    50
# View first few Ensembl IDs
head(rownames(TCGA_LUSC_ENSEMBL))
#> [1] "ENSG00000000003" "ENSG00000000005" "ENSG00000000419" "ENSG00000000457"
#> [5] "ENSG00000000460" "ENSG00000000938"